requirement.txt
dj-database-url==0.5.0
Django==2.2.11
django-countries==6.1.2
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-tinymce==3.0.2
djangorestframework==3.11.0
docutils==0.16
filelock==3.0.12
gunicorn==20.0.4
h5py==2.9.0
python-dateutil==2.8.0
pytz==2018.9
